Skip erd link if none present (#11)
If there is no ERD, skip adding the link to it. There is a separate issue #6 where we want to perhaps find a better solution for hosting the ERD rather than in the `public/` folder, but for now, we could use a fix to prevent a 404 on the link if no ERD is present.

We could just skip adding the `/erd.png` link in the initializer, but that requires us to remember and check that always. A better solution is to handle it in the gem in case we forget to remove it from the initializer. Then we can use a uniform initializer across all of our Rails projects and not need to have different configurations for different projects and have to keep them all up to date.