Build a page with instructions for ERD generation (#13)
The ERD shown at the dev_toolbar link should be synced with the project database. Because resyncing requires running `rake erd`, we need some instructions for regenerating the data model. Also, we should not host the erd in the public folder since that makes it accessible on prduction.

This PR creates an engine and RCAV with instructions for running `rake erd` and showing the `erd.png` file at the root path. Separately, on https://github.com/firstdraft/project-syncing/pull/19, the gem version is bumped and the `bundle exec rails g erd:install` was run to add the auto generating of the ERD to any `rake db:` task (i.e., there will be no need to `rake erd` following a `rake db:migrate`).